requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=8.0.0
matplotlib>=3.3.0
kivy>=2.1.0
pytesseract>=0.3.8
